requirement.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
distlib==0.3.6
filelock==3.12.0
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
platformdirs==3.5.0
virtualenv==20.23.0
Werkzeug==2.3.3
Pillow==9.5.0